fix: Resolve ambiguous call to Math.Round
This commit fixes a build error caused by an ambiguous call to `Math.Round`. The error occurred because the type of the argument could be implicitly converted to both `decimal` and `double`.

The fix ensures that the calculation is performed using floating-point arithmetic, which resolves the ambiguity.

+++ services/FunQuarter.cs
@@ -162,7 +162,7 @@ public static class funscript_converter_funquarter
 
                 if (action.Type != "pause" && lastAction.SubActions.Any())
                 {
-                    var newPos = (lastAction.SubActions.Sum(a => a.pos) + action.pos) / (lastAction.SubActions.Count + 1);
+                    var newPos = (double)(lastAction.SubActions.Sum(a => a.pos) + action.pos) / (lastAction.SubActions.Count + 1);
                     outputAction.pos = (int)Math.Round(newPos);
                     pos = outputAction.pos;
                 }
